requirements.txt
instagrapi
cryptography
pyyaml
rich
pytz
moviepy
certifi
chardet
decorator
idna
imageio
imageio-ffmpeg
numpy
Pillow
pip
proglog
pycryptodomex
pydantic
PySocks
requests
setuptools
tqdm
typing-extensions
urllib3
wheel
python-dotenv ##2fa
langchain==0.2.16
langchain-openai==0.1.25
slack_sdk